Refactor/cleanup and pin aioresponses to 0.6.1
* aioresponses 0.6.2 uses deepcopy which requires all custom objects
  passed to aiohttp request arguments to be picklable.
